services/rag/requirements.txt
Pinecone
pinecone[grpc]>=5.0.0

python-dotenv